# Improve ratestore performance