This PR changes the Rate Store to use multi-level maps rather than `tenant+stream-fingerprint` key. 

Benchmarks:

*old*
```
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/distributor
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkRateStore-8         117          10420080 ns/op         3201761 B/op     200025 allocs/op

PASS
ok      github.com/grafana/loki/pkg/distributor 14.646s
```

*new*
```
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/distributor
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkRateStore-8         320           3811906 ns/op            2660 B/op         26 allocs/op
```

*stats*
```
name         old time/op    new time/op    delta
RateStore-8    10.7ms ± 8%     4.1ms ± 8%  -61.42%  (p=0.000 n=10+10)

name         old alloc/op   new alloc/op   delta
RateStore-8    3.20MB ± 0%    0.00MB ± 0%  -99.92%  (p=0.000 n=8+9)

name         old allocs/op  new allocs/op  delta
RateStore-8      200k ± 0%        0k ± 0%  -99.99%  (p=0.000 n=9+10)
```